moved query params and validator
