+++ x/stakeibc/client/cli/query_validator.go
@@ -1,39 +1,1 @@
 package cli
-
-import (
-	"context"
-
-	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/spf13/cobra"
-
-	"github.com/Stride-Labs/stride/v22/x/stakeibc/types"
-)
-
-func CmdShowValidators() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "show-validators [chain-id]",
-		Short: "shows validators",
-		Args:  cobra.ExactArgs(1),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			clientCtx := client.GetClientContextFromCmd(cmd)
-
-			queryClient := types.NewQueryClient(clientCtx)
-
-			chainId := args[0]
-
-			params := &types.QueryGetValidatorsRequest{ChainId: chainId}
-
-			res, err := queryClient.Validators(context.Background(), params)
-			if err != nil {
-				return err
-			}
-
-			return clientCtx.PrintProto(res)
-		},
-	}
-
-	flags.AddQueryFlagsToCmd(cmd)
-
-	return cmd
-}
